Fix #130611: Grease pencil airbrush affects material of other brushes
This if-statement would change the active unpinned material, even though
it was only meant to handle the pinned material. This was not a new
issue, it was just unlikely to run into it.

Before brush assets were introduced this if-statement would not be
executed when activating any of the default brushes. Afterwards it would
be, because the material was newly linked (because the brush asset was
linked) and so there was no material slot created for it yet.