Fix #132643: Deleting Color Ramp's penultimate stop doesn't update the Color Ramp
This bug occurs when the user deletes the rightmost stop on a color ramp with 2 stops. After doing this, the color ramp appears to stop updating its results.

This bug is due to a shader optimization for color ramps. For ramps with 2 stops or less, the first and second stops are hardcoded as inputs to get the results of the color ramp. When deleting the rightmost stop, the stop's data isn't actually changed. The ramp just updates its total stop number. Because of this, the result of the color ramp won't change.

This is fixed by only allowing this optimization when a ramp has exactly 2 stops. This approach is used since it doesn't make sense to use one stop for a color ramp. Furthermore, this change is more inline with the decision to hardcode the first two stops as inputs for the ramp's results.