Vulkan: Pre-extract pipeline barriers
Pipeline barriers were extracted when recording commands. This works,
but had the downside that it locked the device resources. Extracting
pipeline barriers is fairly small task compared to recording commands.

This PR will perform the extraction of pipelines separate from command
recording. Code is easier to follow and when working with multiple threads
this will reduce locking (more developments are needed).

Original developed in !131965

**TODO**

- Add support for local read