Build the material list for projection painting from the evaluated object.
Texture painting on an object with modifiers that change the materials
(like for example geometry nodes) could lead to unexpected results because
teh material slots were read from the original object, but the eventual
painting was (obviously) done on the evaluated object. This could either
lead to wrong results because the material indices would not match or not
paint at all if the resulting indices were higher than the original number
of slots.

This patch adds a size for the material list to the projection painting state
(instead of relying on Object->totcol) and builds the list of materials from
the evaluated object.