Python Module: Bundle VFX library python bindings
When calling bpy.utils.expose_bundled_modules(), these modules are
added to the sys.path.

This provides a solution/workaround to two problems:
* Using bpy together with packages like usd-core is problematic. Besides
  crashing due to C/C++ symbol conflicts, it's just impossible to import
  different versions of the same module, or to have distinct environment
  variables for both.
* Blender add-ons using these VFX modules do not currently work with
  the bpy module.

A more automatic solution would be to update sys.path on import.bpy,
but perhaps this is too unexpected a side effect. But it remains a
problem that this new mechanism is not easily discoverable.