Fix incorrect reprojection by active smoothing/jitter due to fixed projection plane.
The projection plane changes in stroke projection mode, but active smoothing and
jitter always use the newest projection, even for points "in the past" that are
reprojected. This causes discontinuities because points inside the smoothing
window which should be on interpolated between existing strokes are then projected
onto a different plane.

The solution is to store the projection plane for each point and update it before
applying the active jitter/smoothing retroactively to points inside the window.