Fix: audit and fixes for uses of `Action.idroot`
Did a pass through the code base to make sure the uses of Action.idroot
are correct in the context of layered actions. I found and fixed one
definite issue in grease pencil code where the idroot was being set to a
non-zero value without checking if the action was layered or not first.
Other than that, I also cleaned up a couple of dodgey-looking things I
found.