Compositor: Use BT.709 luminance for color to float conversion
This patch uses BT.709 luminance for implicit conversion from color to
float in the compositor. This is done to match other node systems like
Geometry Nodes and Shader nodes, and because luminance is a much better
default than the average formula used before.

Versioning was added to retain average conversion for old files.