VSE: speedup strip_relations_find_and_invalidate_metas
What the function needs is to invalidate all meta strips that contain
the given strip. It was scanning the whole world to figure that out,
but we already maintain a data structure for finding which meta a
strip is part of.

Updating colors of 4800 strips goes from 270ms to 205ms with this
change (Ryzen 5950X)