Keymap: Add image editor zoom level shortcut on Shift + 1, Ctrl + 1
When zooming in the image editor, one can press 1, 2, 4, 8 on the
numpad to get x1, x2, etc. levels.
One can also do Ctrl/Shift + 2, Ctrl/Shift + 4, etc. to get x0.5,
x0.25 zooms, etc.
However one cannot do Ctrl/Shift + 1 to get back to x1 zoom, which is
frustrating when switching back and forth between zoom levels.

This commit adds shortcuts for Ctrl + 1 and Shift + 1 to get x1 zoom.
In order for the View > Zoom menu to be unchanged, the shortcuts are
sorted so that the modifierless ones come first.
The same change is done to the Industry Compatible keyconfig, on
Shift + F1 and Ctrl + F1.