Fix #113760: Knife Tool: wrong red point after undo
After confirming a cut, the tool goes into `MODE_IDLE`.
In `knifetool_undo`, the `KnifePosData` `prev` (drawn as the "red point"
in question in `knifetool_draw`) was only set to the `KnifeUndoFrame`
pos when we are in `MODE_DRAGGING` though -- this is now changed to also
be done for `MODE_IDLE`.

NOTE: it might be questionable to even draw this bigger red dot once a
cut is confirmed (will attach another small diff to this PR that does
this)