Fix #125192: rigidbody.objects_add discrepancy between exec and poll
Calling `rigidbody.objects_add` fails when a rigid body can’t be added
to the **active** object, even if the active object isn't selected and a
rigid body can be added to all **selected** objects.

`Exec` was working on selected objects, whereas `Poll` was checking the
active object.

To resolve, check the first object in context `selected_objects` [not
all of them, the operator supports skipping non-fitting types], rather
than the active object.