Cycles: oneAPI: use ocloc 101.6557 on Windows
This released version of the graphics compiler allows
the use of some new software functionality on Intel GPUs, such
as SYCL bindless textures support.

Graphics compiler upgrades require increasing the minimum required
driver (compute-runtime) version to the corresponding one to guarantee
compatibility, which is XX.XX.101.6557 in this release, so we bump this
requirement accordingly.