Fix #130106: Vulkan: Pixelbuffer performance
Cycles uses pixel buffers to update the display. Due to making things
work the vulkan backend downloaded the GPU allocated pixel buffer to the
CPU, Copied it to a GPU allocated staging buffer and update the display
texture using the staging buffer. Neatless to say that a GPU->CPU->GPU
roundtrip is not performant.

This PR fixes this by allowing the pixel buffer to act as a staging
buffer as well.

Final image rendering of BMW has similar performance between OpenGL and
Vulkan.
Viewport performance is now also similar. (timings will need to be
measured.