Fix: #133453 Canvas button cannot be enabled
This fixes an issue where the Canvas button is always disabled when trying to use the experimental feature, Sculpt Texture Paint.

This bug was caused by the new Asset System since it changed how paint tools work. Before, the UI would check the current tool being used and determine if the Canvas button should be activated. With the new Asset System, painting is done with modular brushes instead of tools, so previous methods of checking tool properties have stopped working.

This is fixed by activating the Canvas button based on the properties of the current brush. If the brush's type is either Paint or Smear, then the Canvas button is activated.