Fix for T133605: Crash when using VR and an error occurs in the xr context ---
When an error occurred in the xr draw code, wm_xr_error_handler
previously freed the XR context and continued execution, which
would immediately lead to a crash in wm_draw_surface due to a stale
pointer.

This code adds a bool exit_pending to wxXrData, and instead of freeing
the xr context when an error occurs, it now sets this to true, which
allows the wm_draw_surface to complete it's normal work of calling
GPU_context_end_frame() which will still have a valid context pointer
to work with.

On the next loop of WM_main, the wm_xr_events_handle will detect the
exit_pending and will release the context at a time where there
are no further dependencies on it.

https://projects.blender.org/blender/blender/issues/133605