Anim: emit liboverride on slot handle when action is changed
Emit a 'diff' for the `animdata.slot_handle` property whenever the
`animdata.action` property is changed through a library override.

The slot handle is only meaningful within the context of the assigned
action. So when a liboverride changes the assigned action, the slot
handle should also get an override.

This is necessary even when the numerical value of the slot handle
happens to be the same in both actions, as the newly chosen slot is
different from the slot that was chosen in the library file.

-----------

> [!Note]
> Marked as WIP because the same approach is needed in other places
> where slots are assigned as well (NLA strips and Action Constraints).
>
> First I'd like to discuss the approach, though, before I generalise it
> and apply it to the other cases.