Anim: set Action.idroot when writing blend file, for forward compat
When writing an Action to a blend file, for forward compatability
reasons the animation data in the first slot is also written to the old
Animato Action storage locations. That way, loading the Action in an
older version of Blender will still at least have that animation data.

However, we overlooked `Action.idroot`, which indicates which ID
type the Action is for, and therefore all Actions have an unspecified
`idroot` when loaded in pre-layered-action versions of Blender.

This PR fixes that by also setting the `idroot` to match the target ID
type of the first slot when writing to disk.