Anim: change RNA Action.id_root to have backwards-compatible behavior
Most of the old Animato properties on an Action (e.g. FCurve list,
Channel Groups) already act as proxies for the data in the first strip
of the first layer for the first slot.

However, we overlooked `Action.id_root`, which it turns out is writable!

This PR changes `Action.id_root` to act as a proxy for the first Slot's
`target_id_type` property, both for reading and writing.

However, there is no guarantee that there is a Slot. When there are no
Slot's yet, the original `Action.idroot` is still used under the hood.