CMake: Rename and expose 'USE_EXPEIRMENTAL_TESTS' option
The `USE_EXPERIMENTAL_TESTS` variable was not exposed as an option, so
the two tests that use the option to check whether tests should be run
had to be manually enabled by changing `tests/python/CMakeLists.txt`.

This commit renames the variable to `WITH_EXPERIMENTAL_TESTS`, defaults
the option to `OFF`, and marks it as an advanced option.